+++ src/main/java/com/enmusubi/invitation/InvitationController.java
@@ -6,13 +6,24 @@ import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 
 @WebServlet("/InvitationController")
 public class InvitationController extends HttpServlet {
     private static final long serialVersionUID = 1L;
 
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        InvitationDAO.invitations(request);
+    	HttpSession session = request.getSession(false);
+
+		if (session != null && session.getAttribute("m_name") != null) {
+			System.out.println("세션 생존");
+			request.setAttribute("mainNav", "mainNavAF.jsp");
+		} else {
+			System.out.println("세션 죽음");
+			request.setAttribute("mainNav", "mainNavBF.jsp");
+		}
+    	
+    	InvitationDAO.invitations(request);
         request.getRequestDispatcher("invitationPage/invitationPage.jsp").forward(request, response);
     }
