+++ src/main/webapp/invitationPage/mainNavBF.jsp
@@ -1,0 +1,36 @@
+<%@ page language="java" contentType="text/html; charset=UTF-8"
+	pageEncoding="UTF-8"%>
+<!DOCTYPE html>
+<html>
+<div class="yj-main-menubox">
+	<div class="yj-main-menu">
+		<img alt="" src="main/imgFolder/yj-main-menu1.png" />
+	</div>
+	<div class="yj-main-nav nav-hidden">
+		<div class="nav-fakeDiv"></div>
+		<div class="nav-logoDeco">
+		<img src="main/imgFolder/yj-main-logo.png" alt="nav-logo">
+		</div>
+		<div class="nav-login">
+			<a href="MemberC"><h3>ログイン</h3></a>
+		</div>
+		<div class="nav-register">
+			<a href="MemberRegC"><h3>会員登録</h3></a>
+		</div>
+		<div class="nav-productBF">
+			<a href="ProductController"><h3>商品紹介</h3></a>
+		</div>
+		<div class="nav-text">
+			<h3 class="nav-textDeco">✿✿✿</h3>
+			<h3>
+				ログインして <br /> より多くのサービスを <br /> ご利用ください
+			</h3>
+			<h3 class="nav-textDeco">✿✿✿</h3>
+		</div>
+		<div class="nav-topBtn">
+			<a href="#"><img
+				src="main/imgFolder/yj-main-topBtn.png" alt="nav-topBtn" /></a>
+		</div>
+	</div>
+</div>
+</html>
