--- src/main/webapp/mytemplate/css/mytemplate.css
@@ -1,294 +1,0 @@
-/* Google Fonts에서 폰트 가져오기 */
-@import url('https://fonts.googleapis.com/css2?family=Noto+Serif+JP:wght@200..900&display=swap');
-
-/* 공통 스타일 */
-body {
-    font-family: "Noto Serif JP", serif;
-    font-optical-sizing: auto;
-    font-weight: 400;
-    font-style: normal;
-    margin: 0;
-    padding: 0;
-    position: relative;
-    display: flex;
-    flex-direction: column;
-    background-color: #FFF5F4;
-   background-image: url('imgFolder/upimg.png'), url('imgFolder/downimg.png');
-}
-
-/* 헤더 스타일 */
-.hw_head {
-    background-position: top center;
-    background-attachment: scroll;
-    background-repeat: no-repeat;
-    padding: 2rem;
-}
-
-.hw_logo {
-    z-index: 4;
-    width: 150px;
-    height: 150px;
-    position: absolute;
-    top: 3rem;
-    left: 4rem;
-}
-
-.hw_logo img {
-    width: 100%;
-    height: 100%;
-    animation: logo-float 2s ease-in-out infinite;
-}
-
-.hw_page_title {
-    position: absolute;
-    left: 50%;
-    transform: translateX(-50%);
-    top: 70px;
-}
-
-@keyframes logo-float {
-    0%, 100% {
-        transform: translate(0, 0%);
-    }
-    50% {
-        transform: translate(0, 5%);
-    }
-}
-
-.hw_menu {
-    z-index: 4;
-    width: 80px;
-    height: 80px;
-    position: absolute;
-    top: 3rem;
-    right: 5rem;
-}
-
-.hw_menu img {
-    width: 100%;
-    height: 100%;
-}
-
-/* 페이지 컨테이너 */
-.hw_container {
-    border-radius: 1rem;
-    max-width: 90%;
-    margin: auto;
-    position: relative;
-    z-index: 1;
-    padding: 0 1rem;
-}
-
-/* 네비게이션 바 스타일 */
-.hw_menubar {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    margin-top: 15rem;
-    padding: 3rem;
-}
-
-.hw_menubar_item {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    margin-bottom: 1rem;
-    padding: 0.5rem;
-}
-
-.hw_menubar_menu {
-    font-size: 1.25rem;
-    text-decoration-line: none;
-    color: #696969;
-    transition: color 0.3s ease;
-}
-
-.hw_menubar_menu:hover {
-    color: #ff2828;
-}
-
-.hw_menubar_menu.active {
-    color: #ff2828;
-}
-
-.hw_yellow_line {
-    width: 9rem;
-    height: 1rem;
-}
-
-/* 메인 스타일 */
-main {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    width: 100%;
-}
-
-/* 버튼 스타일 */
-.top_button {
-    display: block;
-    margin: 0 auto 10rem auto;
-    width: 8.7rem;
-    height: 8.7rem;
-    cursor: pointer;
-    padding: 1rem;
-}
-
-/* 숨김/표시 클래스 */
-.hw-f-block {
-    display: block;
-}
-
-.hw-f-none {
-    display: none;
-}
-
-/* 제품 페이지 컨텐츠 */
-.hw_product-page-content {
-    margin-top: 4.3rem;
-}
-
-/* 페이지 컨텐츠 */
-.hw_page-content {
-    display: grid;
-    grid-template-columns: repeat(3, 1fr);
-    gap: 1rem;
-}
-
-.hw_page-content-template {
-    padding: 1rem;
-    border-radius: 0.5rem;
-    position: relative;
-    overflow: hidden;
-}
-
-.hw_template-button {
-    position: absolute;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%);
-    display: flex;
-    justify-content: center;
-    padding: 0.5rem;
-    gap: 2rem;
-    z-index: 2;
-    flex-direction: column;
-    opacity: 0;
-    padding: 15rem;
-    border-radius: 30%;
-}
-
-.hw_template-button:hover {
-    opacity: 1;
-}
-
-.hw_template-button > button {
-    width: 7rem;
-    height: 1.8rem;
-    font-size: 1rem;
-    font-weight: bold;
-    border: 0.1rem solid #c82626;
-    border-radius: 0.5rem;
-    box-shadow: 0.1rem 0.1rem 0.3rem #c82626;
-    z-index: 100;
-    background-color: #fff4f4;
-    color: #5c5c5c;
-}
-
-.hw_template-img {
-    display: flex;
-    position: relative;
-    justify-content: center;
-    align-items: center;
-}
-
-.hw_template-img > img {
-    width: 100%;
-}
-
-.hw_template-title {
-    text-align: center;
-    margin-top: 1.5rem;
-    font-size: 1.3rem;
-    font-weight: bold;
-}
-
-.hw_share_btn {
-    background: none;
-    border: none;
-    margin: 0.5rem auto;
-    cursor: pointer;
-    display: block;
-    width: 7rem;
-    height: 1.8rem;
-    font-size: 1rem;
-    font-weight: bold;
-    border: 0.1rem solid #c82626;
-    border-radius: 0.5rem;
-    box-shadow: 0.1rem 0.1rem 0.3rem #c82626;
-    background-color: #fff4f4;
-    color: #5c5c5c;
-    transition: background-color 0.5s, color 0.5s;
-}
-
-.hw_share_btn:hover {
-    background-color: #ff3b3f;
-    color: #fff4f4;
-}
-
-/* 페이지네이션 */
-.pagination {
-    text-decoration-line: none;
-    margin-top: 8%;
-}
-
-/* 팝업 창 스타일 */
-.popup {
-    display: none;
-    position: fixed;
-    z-index: 3;
-    left: 50%;
-    top: 50%;
-    transform: translate(-50%, -50%);
-    width: 300px;
-    padding: 20px;
-    box-shadow: 0 5px 15px rgba(0, 0, 0, 0.3);
-    background-color: white;
-    border-radius: 10px;
-}
-
-.popup-content {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-}
-
-.popup-content input {
-    width: 100%;
-    padding: 10px;
-    margin-bottom: 10px;
-    border: 1px solid #ddd;
-    border-radius: 5px;
-}
-
-.popup-content button {
-    padding: 10px 20px;
-    background-color: #007bff;
-    color: white;
-    border: none;
-    border-radius: 5px;
-    cursor: pointer;
-}
-
-.popup-content button:hover {
-    background-color: #0056b3;
-}
-
-.close-btn {
-    position: absolute;
-    top: 10px;
-    right: 10px;
-    font-size: 20px;
-    cursor: pointer;
-}
